+++ CHANGES.md
@@ -1,4 +1,8 @@
-# 0.3.0 / 2021-12/22
+# 0.3.1 / 2021-12-24
+
+  * Deprecation warning that fasttext will be removed
+
+# 0.3.0 / 2021-12-22
 
   * Sentence augementations
   * Demo notebooks added